Added Jump and Fixed Camera
Changed script to all Vector3 inputs. This allows for the jump functionality which is called and instantiated through the same input as moving, but on the Y axis. Changed BallController to apply the BallSpeed only to the Z and X components, and added a JumpSpeed that applies to the Y component.